Conflict 409 on MRD delete if MRD is associated to any task

+++ src/main/java/com/ef/mediaroutingengine/services/controllerservices/MediaRoutingDomainsServiceImpl.java
@@ -1,13 +1,21 @@
 package com.ef.mediaroutingengine.services.controllerservices;
 
+import com.ef.mediaroutingengine.dto.MrdDeleteConflictResponse;
+import com.ef.mediaroutingengine.dto.SuccessResponseBody;
 import com.ef.mediaroutingengine.exceptions.NotFoundException;
 import com.ef.mediaroutingengine.model.MediaRoutingDomain;
 import com.ef.mediaroutingengine.model.PrecisionQueueEntity;
+import com.ef.mediaroutingengine.model.Task;
 import com.ef.mediaroutingengine.repositories.MediaRoutingDomainRepository;
 import com.ef.mediaroutingengine.repositories.PrecisionQueueEntityRepository;
+import com.ef.mediaroutingengine.services.pools.TasksPool;
 import java.util.List;
 import java.util.UUID;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,7 +24,10 @@ import org.springframework.transaction.annotation.Transactional;
  */
 @Service
 public class MediaRoutingDomainsServiceImpl implements MediaRoutingDomainsService {
-
+    /**
+     * The constant LOGGER.
+     */
+    private static final Logger LOGGER = LoggerFactory.getLogger(MediaRoutingDomainsServiceImpl.class);
     /**
      * The Repository.
      */
@@ -26,6 +37,8 @@ public class MediaRoutingDomainsServiceImpl implements MediaRoutingDomainsServic
      */
     private final PrecisionQueueEntityRepository precisionQueueEntityRepository;
 
+    private final TasksPool tasksPool;
+
     /**
      * Constructor, Autowired, loads the beans.
      *
@@ -34,9 +47,11 @@ public class MediaRoutingDomainsServiceImpl implements MediaRoutingDomainsServic
      */
     @Autowired
     public MediaRoutingDomainsServiceImpl(MediaRoutingDomainRepository repository,
-                                          PrecisionQueueEntityRepository precisionQueueEntityRepository) {
+                                          PrecisionQueueEntityRepository precisionQueueEntityRepository,
+                                          TasksPool tasksPool) {
         this.repository = repository;
         this.precisionQueueEntityRepository = precisionQueueEntityRepository;
+        this.tasksPool = tasksPool;
     }
 
     @Override
@@ -63,15 +78,18 @@ public class MediaRoutingDomainsServiceImpl implements MediaRoutingDomainsServic
 
     @Override
     @Transactional
-    public List<PrecisionQueueEntity> delete(UUID id) {
+    public ResponseEntity<Object> delete(UUID id) {
         if (!this.repository.existsById(id)) {
             throw new NotFoundException("Could not find the resource to delete");
         }
         List<PrecisionQueueEntity> precisionQueueEntities = this.precisionQueueEntityRepository.findByMrdId(id);
-        if (precisionQueueEntities.isEmpty()) {
+        List<Task> tasks = this.tasksPool.findByMrdId(id);
+        if (precisionQueueEntities.isEmpty() && tasks.isEmpty()) {
             this.repository.deleteById(id);
+            return new ResponseEntity<>(new SuccessResponseBody("Successfully Deleted"), HttpStatus.OK);
         }
-        return precisionQueueEntities;
+        LOGGER.debug("Could not delete MRD: {}. It is associated with one or more Queues", id);
+        return new ResponseEntity<>(new MrdDeleteConflictResponse(precisionQueueEntities, tasks), HttpStatus.CONFLICT);
     }
 
     /**
